requirements.txt
streamlit
langchain_groq
langchain-openai
python-dotenv
langchain
langchain_core
openai
flask
flask-cors